requirements.txt
fastapi>=0.95.0
uvicorn>=0.22.0
requests>=2.28.2
pyyaml>=6.0
pydantic>=2.7.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0